Create AbstractStructureFilter
One way to optimize this code is to use Python's built-in caching mechanism. We can cache the results of the test method for each structure, so that if the same structure is tested multiple times, we don't have to perform the test again.

To implement this, we can use the functools module, which provides the lru_cache decorator. This decorator will automatically cache the results of the test method for us.

By adding the @lru_cache(maxsize=None) decorator to the test method, we enable caching with an unlimited cache size. This means that the results of the method will be cached for all structures that are tested, and if the same structure is tested again, the cached result will be returned instead of performing the test again.

This optimization can significantly improve the performance of the code, especially if the test method is computationally expensive.

Signed-off-by: hamdi gokhan ince <66874195+Hamdi57@users.noreply.github.com>